Store player ID as long instead of int

+++ app/src/main/java/io/github/fobo66/wearmmr/api/MatchmakingRatingApi.kt
@@ -28,5 +28,5 @@ import retrofit2.http.Path
 interface MatchmakingRatingApi {
 
     @GET("players/{id}")
-    fun fetchPlayerProfile(@Path("id") id: Int): Observable<PlayerInfo>
+    fun fetchPlayerProfile(@Path("id") id: Long): Observable<PlayerInfo>
 }
